Adding in stories.
Added in the first set of stories (basic), and starting in on
management.

Ordering is controlled in _config.

Includes clean up/regularize the inclusion of chartjs. There are other
ways, but this works well enough. I experimented with some other
approaches, and ultimately, every vis is going to be rather customized.

Yes, there was probably a full-on OO or React-based way to do this. I'm
interested in telling a story, and expect it to be told once. Fool me if
this lives for 20 years.